requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.120.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
starlette==0.49.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0

aioredis~=2.0.1
celery~=5.3.6